requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
pandas-datareader==0.9.0
plotly==4.13.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
retrying==1.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
